refactor: editors decoration
