+++ constant/media.js
@@ -1,3 +1,5 @@
 export const MEDIA = {
   HTML: "HTML",
+  PLANE: "PLANE",
+  BBOX: "BBOX",
 };
